refactor: moved computeType into appropriate folder
