+++ packages/gatsby-transformer-raml/src/utils/compute-type.js
